+++ pylabianca/stats.py
@@ -40,6 +40,10 @@ def permutation_test(*arrays, paired=False, n_perm=1000, progress=False,
     stat_fun = borsar.stats._find_stat_fun(n_groups=n_groups, paired=paired,
                                             tail=tail)
 
+    has_xarr = all(['DataArray' in str(type(x)) for x in arrays])
+    if has_xarr:
+        arrays = [x.values for x in arrays]
+
     thresh, dist = borsar.stats._compute_threshold_via_permutations(
         arrays, paired=paired, tail=tail, stat_fun=stat_fun,
         return_distribution=True, n_permutations=n_perm, progress=progress,
@@ -134,7 +138,7 @@ def cluster_based_test(frate, compare='image', cluster_entry_pval=0.05,
     Returns
     -------
     stats : numpy.ndarray
-        Anova F statistics for every timepoint.
+        Anova F statistics for every time point.
     clusters : list of numpy.ndarray
         List of cluster memberships.
     pval : numpy.ndarray
@@ -161,7 +165,7 @@ def cluster_based_test(frate, compare='image', cluster_entry_pval=0.05,
 # ENH: move to sarna/borsar sometime
 # ENH: allow for standard arrays (and perm_index)
 def cluster_based_test_from_permutations(data, perm_data, tail='both',
-                                         adjacency=None):
+                                         adjacency=None, percentile=5):
     '''Performs a cluster-based test from precalculated permutations.
 
     This function should get data ready for cluster-based permutation test
@@ -185,6 +189,16 @@ def cluster_based_test_from_permutations(data, perm_data, tail='both',
     adjacency : numpy.ndarray
         Adjacency matrix for clustering. If ``None`` then lattice adjacency is
         used.
+    percentile : int or float
+        Percentile of the permutation distribution to use for thresholding.
+        The percentile threshold is calculated for every data point (time point
+        for example) separately. The percentile conforms to the tail selected
+        using the ``tail`` parameter. When ``tail`` is ``'both'`` then two
+        percentiles are calculated (one for positive and one for negative tail
+        ). For example, if ``percentile`` is 5, and the tail is ``'both'`` then
+        the 2.5th (``percentile / 2``) and 97.5th (``100 - (percentile / 2)``)
+        percentiles are used as thresholds. The percentile should be between
+        0 and 100.
 
     Returns
     -------
@@ -202,20 +216,20 @@ def cluster_based_test_from_permutations(data, perm_data, tail='both',
     assert isinstance(perm_data, xr.DataArray)
     assert tail in ['both', 'pos', 'neg']
     dim_names = ['perm', 'permutation']
-    has_dim = [dimname in perm_data.dims for dimname in dim_names]
+    has_dim = [dim_name in perm_data.dims for dim_name in dim_names]
     assert any(has_dim)
     perm_dim_name = dim_names[np.where(has_dim)[0][0]]
-
-    percentiles = [97.5, 2.5] if tail == 'both' else [95, 5]
     perm_dim = perm_data.dims.index(perm_dim_name)
-    n_perms = perm_data.shape[perm_dim]
-    thresholds = np.percentile(perm_data, percentiles, axis=perm_dim)
 
     if tail == 'both':
+        percentiles = [100 - (percentile / 2), (percentile / 2)]
+        thresholds = np.percentile(perm_data, percentiles, axis=perm_dim)
         thresholds = [thresholds[0], thresholds[1]]
     elif tail == 'pos':
-        thresholds = [thresholds[0], None]
+        thresholds = np.percentile(perm_data, 100 - percentile, axis=perm_dim)
+        thresholds = [thresholds, None]
     elif tail == 'neg':
+        thresholds = np.percentile(perm_data, percentile, axis=perm_dim)
         thresholds = [None, thresholds[1]]
 
     # clusters on actual data